refactor: 범용적인 로직 디렉토리 경로 수정
